# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentFlow Intelligence",
    "description": "Agent automated evaluation platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tenants": {
      "get": {
        "tags": [
          "多租户"
        ],
        "summary": "List Tenants",
        "description": "List tenants visible to the caller.\n\nsystem_admin: all tenants.\nothers: memberships only.",
        "operationId": "list_tenants_api_v1_tenants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Tenants Api V1 Tenants Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "多租户"
        ],
        "summary": "Create Tenant",
        "description": "Create a tenant; creator becomes tenant_admin member.",
        "operationId": "create_tenant_api_v1_tenants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Tenant Api V1 Tenants Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}": {
      "get": {
        "tags": [
          "多租户"
        ],
        "summary": "Get Tenant",
        "description": "Get tenant if caller is member or system_admin.",
        "operationId": "get_tenant_api_v1_tenants__tenant_id__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Tenant Api V1 Tenants  Tenant Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_id}/members": {
      "post": {
        "tags": [
          "多租户"
        ],
        "summary": "Add Member",
        "description": "Add or update a tenant member (tenant_admin / system_admin).",
        "operationId": "add_member_api_v1_tenants__tenant_id__members_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantMemberAdd"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Add Member Api V1 Tenants  Tenant Id  Members Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "多租户"
        ],
        "summary": "List Members",
        "operationId": "list_members_api_v1_tenants__tenant_id__members_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Members Api V1 Tenants  Tenant Id  Members Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TenantCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Slug"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "title": "TenantCreate"
      },
      "TenantMemberAdd": {
        "properties": {
          "user_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "maxLength": 32,
            "title": "Role",
            "default": "member"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "TenantMemberAdd"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}